create default dataset and baskets on schema import with wizard
